CR: Extract setupTestHooks
