+++ local-proxy/package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "build": "rm -rf dist/ && tsc",
     "lint": "tslint -c ../common/tslint.yml -p .",
-    "test": "ava -v dist/test/*.js"
+    "test": "ava -v dist/test/*.test.js"
   },
   "author": "",
   "license": "MIT",
